+++ src/shifttrade/pages/2_Solicitud.py
@@ -2,10 +2,10 @@ import streamlit as st
 from datetime import datetime
 
 try:
-    import utils
-    import supabase_client
-    import token_utils
-    import email_utils
+    from shifttrade import utils
+    from shifttrade import supabase_client
+    from shifttrade import token_utils
+    from shifttrade import email_utils
 except st.errors.StreamlitSecretNotFoundError as e:
     st.error(
         "CRITICAL ERROR: Could not load application secrets required by 'utils.py'.\n"
